feat: Add function to make grid and static color
Static color is on hover only, not click. I would ideally like to make
it a click and drag situation. Or a toggle situation (on click, hover to
color. On click again, hover to color stops).

Next step: add button to top of screen asking user for grid size, max
100x100.