feat: make cells blue upon user selected size
Previously, I couldn't get the mouseover blue to work on user input grid
sizes because the cells array was empty. So I stuck the cells array
inside the function that makes the grid and that fixed the problem.

Next: see if I can make a makeBlue function outside the makeGrid
function and call the makeBlue function inside makeGrid.

Also need to work on the look of the page and how to make different
color options, like random colors.